requirements.txt
flask 
pillow
werkzeug